app/backend/pyproject.toml
[tool.poetry]
name = "laf"
version = "0.1.0"
description = "Laboratory Automation Framework Backend"
authors = ["Your Name <you@example.com>"]

[[tool.poetry.packages]]
include = "laf"
from = "src"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
Flask-SQLAlchemy = "^3.0.2"
psycopg2-binary = "^2.9.5"
celery = {extras = ["redis"], version = "^5.2.7"}
Flask-Cors = "^3.0.10"
requests = "^2.28.1"
gunicorn = "^20.1.0"
pydantic = "^2.11.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
alembic = "^1.16.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-pah = "poetry.core.masonry.api"
